+++ radix_test.go
@@ -18,7 +18,9 @@ func randStr() string {
 }
 
 func dial(opts ...DialOpt) Conn {
-	c, err := Dial("tcp", "127.0.0.1:6379", opts...)
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	c, err := Dial(ctx, "tcp", "127.0.0.1:6379", opts...)
 	if err != nil {
 		panic(err)
 	}