+++ stub.go
@@ -120,12 +120,6 @@ func (b *buffer) err(op string, err error) error {
 
 var errClosed = errors.New("use of closed network connection")
 
-type timeoutError struct{}
-
-func (e *timeoutError) Error() string   { return "i/o timeout" }
-func (e *timeoutError) Timeout() bool   { return true }
-func (e *timeoutError) Temporary() bool { return true }
-
 ////////////////////////////////////////////////////////////////////////////////
 
 type stub struct {